% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jpg_to_png}
\alias{jpg_to_png}
\title{Convert RGB jpg images to 8 bit grayscale png}
\usage{
jpg_to_png(input, outputdir, create.dir = FALSE, lazy = TRUE)
}
\description{
Convert RGB jpg images to 8 bit grayscale png
}
