% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindtct.R
\name{matchscores}
\alias{matchscores}
\title{Compute fingerprint matching scores}
\usage{
matchscores(mindtct_out_A, mindtct_out_B, outputdir = ".", options = "")
}
\arguments{
\item{outputdir}{directory where scores and configuration files are saved.}

\item{options}{options and flags for the \code{bozorth3} program.}

\item{mindtct_out_A}{tibble with a column named "out" containing paths to \code{mindtct} output directories.}

\item{mindtct_out_B}{tibble with a column named "out" containing paths to \code{mindtct} output directories.}
}
\description{
Compute pairwise fingerprint matching scores using \code{bozorth3}.
}
