% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindtct.R
\name{mindtct}
\alias{mindtct}
\title{Minutia Detection}
\usage{
mindtct(imgfiles, outputdir = ".", silent = FALSE, options = "", mc.cores = parallel::detectCores())
}
\arguments{
\item{imgfiles}{list or character vector of paths to image files compatible with \code{mindtct}.}

\item{outputdir}{directory where mindtct output is saved. For each image, a directory named after the image is created within the output directory. Ouput files of \code{mindtct} are saved as \verb{out.*} in their corresponding directory.}

\item{options}{flags and options to pass to the \code{mindtct} program.}

\item{silent}{whether or not to hide messages. Default is FALSE.}

\item{mc.cores}{Optional number of cores accross which to parallelize execution. Default is 1 (no parallelization).}
}
\value{
tibble with one row for each processed image. The column "imagefile" contains the path to the original image file. The column "out" contains the directory path to \code{mindtct} output.
}
\description{
Detect minutiae in fingerprint images using NIST's MINDTCT program.
}
\note{
The \code{mindtct} program, part of NIST Biometric Image Software (NBIS), should be installed and available in PATH. If not in PATH, the path to NBIS executables should be specified as a global option through \verb{options(NBIS_bin = <path_to_binaries>)}.
}
