% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindtct.R
\name{plotMinutiae}
\alias{plotMinutiae}
\title{Plot detected minutiae}
\usage{
plotMinutiae(mindtct_out, maxPlots = 5, col = 3, lwd = 1.5)
}
\arguments{
\item{mindtct_out}{output from the \code{mindtct()} function. This should be a tibble with one row for each processed image. The column "imgfile" should contain the path to the original image. The column "out" should contain the path to the corresponding \code{mindtct} output directory.}

\item{maxPlots}{maximal number of fingerprint images to show (number of rows in the grid plot.)}
}
\description{
Plot fingerprint image, its binarization, and detected minutiae with orientaiton and quality indicator.
}
