% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_fmr}
\alias{print_fmr}
\title{Print fingerprint minutiae record}
\usage{
printFMR(filepath, type)
}
\arguments{
\item{filepath}{character path to file containing the fingerprint minutiae
record (e.g. a "*.ist" file).}

\item{type}{character description of the record standard; one of the following:
\itemize{
\item "ANSI_2004": ANSI/INCITS 378-2004 format
\item "ANSI_2007": ANSI/INCITS 378-2007 format
\item "ISO_2005": ISO/IEC 19794-2:2005 format
\item "ISONC_2005": ISO/IEC 19794-2:2005 normal card format
\item "ISOCC_2005": ISO/IEC 19794-2:2005 compact card format
}}
}
\description{
Print fingerprint minutiae record as specified in ANSI/INCITS 378-2004 and
ISO/IEC 19794-2:2005. Note that ANSI/INCITS 378-2004 and ISO/IEC 19794-2:2005
are older standards which may not be compatible with recent corrections.
}
\note{
This R package is based on \href{https://www.nist.gov/services-resources/software/biomdi-software-tools-supporting-standard-biometric-data-interchange}{NIST's BiomDI software} and includes the
libfmr C library. This function is adapted from the prfmr program.
}
