% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_fmr_raw}
\alias{read_fmr_raw}
\title{Read fingerprint minutiae records}
\usage{
readFMR(filepath, type)
}
\arguments{
\item{filepath}{character path to file containing the fingerprint minutiae
record (e.g. a "*.ist" file).}

\item{type}{character description of the record standard; one of the following:
\itemize{
\item "ANSI_2004": ANSI/INCITS 378-2004 format
\item "ANSI_2007": ANSI/INCITS 378-2007 format
\item "ISO_2005": ISO/IEC 19794-2:2005 format
\item "ISONC_2005": ISO/IEC 19794-2:2005 normal card format
\item "ISOCC_2005": ISO/IEC 19794-2:2005 compact card format
}}
}
\value{
fingerprint minutiae record as a nested list with the following elements:
\itemize{
\item format_std
\item product_identifier_owner
\item product_identifier_type
\item scanner_id
\item compliance
\item x_image_size
\item y_image_size
\item x_resolution
\item y_resolution
\item num_views
\item finger_views: list of fingerprint views each structured as follows:
\itemize{
\item format_std
\item finger_number
\item view_number
\item impression_type
\item finger_quality
\item number_of_minutiae
\item x_image_size
\item y_image_size
\item x_resolution
\item y_resolution
\item algorithm_id
\item minutiae_data: list of minutiae data records each structured as follows:
\itemize{
\item format_std
\item index
\item type
\item x_coord
\item y_coord
\item angle
\item quality
}
}
}
}
\description{
Read fingerprint minutiae records as specified in ANSI/INCITS 378-2004 and
ISO/IEC 19794-2:2005. The record is returned as a nested list structured as described
in the details section. Note that ANSI/INCITS 378-2004 and ISO/IEC 19794-2:2005
are older standards which may not be compatible with recent corrections.
}
\note{
This R package is based on \href{https://www.nist.gov/services-resources/software/biomdi-software-tools-supporting-standard-biometric-data-interchange}{NIST's BiomDI software} and includes the
libfmr C library. This function is adapted from the prfmr program.
}
