% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindtct.R
\name{tidyMinutiae}
\alias{tidyMinutiae}
\title{Extract minutiae information from \code{mindtct} output}
\usage{
function(mindtct_out)
}
\arguments{
\item{mindtct_out}{output from the \code{mindtct()} function. This should be a tibble with one row for each processed image. The column "imgfile" should contain the path to the original image. The column "out" should contain the path to the corresponding \code{mindtct} output directory.}
}
\description{
Construct a tibble containing minutiae coordinates, orientation, and quality.
}
