% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindtct.R
\name{matchscores}
\alias{matchscores}
\title{Compute fingerprint matching scores}
\usage{
matchscores(out_probes, out_gallery, outputdir = ".", options = "")
}
\arguments{
\item{out_probes}{tibble with a column named "out" containing paths to \code{mindtct} output directories.}

\item{out_gallery}{tibble with a column named "out" containing paths to \code{mindtct} output directories.}

\item{outputdir}{directory where scores and configuration files are saved.}

\item{options}{options and flags for the \code{bozorth3} program.}
}
\description{
Compute pairwise fingerprint matching scores using \code{bozorth3}.
}
